# admin/openapi.yaml
openapi: 3.0.2
servers:
  - url: /v1
info:
  version: 1.0.0
  title: Admin space API
...
paths:
...
  /admin/v1/api/blog:
    get:
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        'default':
          $ref: '#/components/responses/Default'
    patch:
      requestBody:
        description: Create a new post
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blog'
      responses:
        '200':
          description: Successful operation
        '401':
          $ref: '#/components/responses/Unauthorized'
        'default':
          $ref: '#/components/responses/Default'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error_message:
          type: string
    Blog:
      type: object
      properties:
        Url:
          type: string
        Title:
          type: string
        Description:
          type: string
        Logo:
          type: string
        Cover:
          type: string
        Themes:
          type: array
          items:
            type: string
        ActiveTheme:
          type: string
        PostsPerPage:
          type: integer
          format: int64
        NavigationItems:
          type: array
          items:
            $ref: '#/components/schemas/Navigation'
    Navigation:
      type: object
      properties:
        label:
          type: string
        url:
          type: string
...
  responses:
    Unauthorized:
      description: Authentication information is missing or invalid
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
security:
  - cookieAuth: [ ]